backend: refactoring SpdxService: moved package and file processing to PackageHandler.java
